add abstraction for promise extension building
